Remove hard-coded BS .active class --> let JS handle .active class...
Having a hard coded .active class on the home link causes the highlighting underline to show momentarily on home for the page-load of any page.
It shows momentarily because the JS hasn't had the chance to strip it out.
Allowing the JS to handle adding the active class eliminates this repaint of the navigation's red underline highlightibg the current navigation location.