Add redirect rule for `events/giveday` & `/giveday`...
Redirect rule will send the following two URLs to https://foundation.kcc.edu/events/annual-fundraiser/:
- https://foundation.kcc.edu/giveday/
- https://foundation.kcc.edu/events/giveday/

Creating a new page for every fundraiser, and adjusting all the links to that page, is exactly
what we were trying to avoid when we redesigned the foundation website.

We had an agreement, throughout the redesign of foundation, that the annual-fundraiser page (https://foundation.kcc.edu/events/annual-fundraiser/)
was to house the content for each year's fundraising event.

In order to keep our agreemen, and avoid unecessary extra work, I am placing the Give Day content on the annual-fundraiser page.
To ensure the link which foundation has already sent-out in print--prior to asking or informing us--will still work, I added another redirect
rule to our web.config file.