Fix fonts that depended on GIVE DAY's Title...
When changing all occurrences of "Give Day" to uppercase, I also changed
the value which gets injected into the title tag of the doc's head.

In the file `_includes/styles/fonts.html`, The conditional liquid statement that
inject's the font-link if the page is Give Day had to be adjusted for "GIVE DAY" instead.