Add contacts collection to CloudCannon `site._explore.collections` listing...
It appears that if a collection (like the contacts collection) is omitted
from the CloudCannon `_explore` configuration (in `_config.yml`) it won't
exist in CloudCannon.

This means that if a site uses custom `_explore` CloudCannon settings in
its Jekyll config file, it must include all collections that editors/admins
need access to for editing.